Revert "split out reference class to it's own file to avoid circular import"
This reverts commit 45e66e01c9236dc5bf00a106f37793753686a59c.
