--- src/reference.js
@@ -1,6 +1,0 @@
-export class Reference {
-  /* jshint unused: false */
-  resolve() {
-    throw new Error('implement resolve in child class');
-  }
-}
